Add a "who's here?" section
I've added a "card" for 3 current Space4 members and cribbed some
details from their respective webpages. I've put a placeholder image
of the correct ratio in for now - although we may want to change
the size later.

It's possible that tiles (http://bulma.io/documentation/layout/tiles/)
are a better layout for this than columns. Each member is likely to
have different text and it may flow oddly in this 3-column layout.